Revert "[SW-1207] Implement H2OTargetEncoder for Scala API (#1192)"
This reverts commit c9aa942c6d16b8ee15a15142620ef17230402a95.
