+++ core/src/test/scala/org/apache/spark/h2o/utils/TestFrameUtils.scala
@@ -91,8 +91,6 @@ object TestFrameUtils extends Matchers {
   }
 
   def assertDataFramesAreIdentical(expected: DataFrame, produced: DataFrame): Unit = {
-    expected.cache()
-    produced.cache()
     val expectedCount = expected.count()
     val producedCount = produced.count()
     assert(
