+++ ml/src/main/scala/org/apache/spark/ml/h2o/models/H2OMOJOPipelineModel.scala
@@ -123,7 +123,7 @@ class H2OMOJOPipelineModel(override val uid: String) extends H2OMOJOModelBase[H2
   }
 
   override def transform(dataset: Dataset[_]): DataFrame = {
-    val frameWithPredictions = applyPredictionUdf(dataset, modelUdf)
+    val frameWithPredictions = applyPredictionUdf(getPredictionCol(), dataset, modelUdf)
 
     val fr = if (getNamedMojoOutputColumns()) {
